# openapi_snapshot.json
{"openapi":"3.1.0","info":{"title":"Healthcare API","description":"Free REST API serving 55,500 synthetic patient encounters. No authentication required. Perfect for demos, learning, and AI agents!","version":"1.0.0"},"paths":{"/":{"get":{"summary":"Root","description":"API root - welcome message and endpoints","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/encounters":{"get":{"summary":"Get Encounters","description":"Get patient encounters with filtering, pagination, and sorting\n\nExample: /api/encounters?condition=Diabetes&min_age=50&limit=20","operationId":"get_encounters_api_encounters_get","parameters":[{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":1000,"minimum":1,"description":"Number of records to return","default":10,"title":"Limit"},"description":"Number of records to return"},{"name":"offset","in":"query","required":false,"schema":{"type":"integer","minimum":0,"description":"Number of records to skip","default":0,"title":"Offset"},"description":"Number of records to skip"},{"name":"condition","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Filter by medical condition","title":"Condition"},"description":"Filter by medical condition"},{"name":"admission_type","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Filter by admission type (Emergency/Urgent/Elective)","title":"Admission Type"},"description":"Filter by admission type (Emergency/Urgent/Elective)"},{"name":"min_age","in":"query","required":false,"schema":{"anyOf":[{"type":"integer","maximum":120,"minimum":0},{"type":"null"}],"description":"Minimum patient age","title":"Min Age"},"description":"Minimum patient age"},{"name":"max_age","in":"query","required":false,"schema":{"anyOf":[{"type":"integer","maximum":120,"minimum":0},{"type":"null"}],"description":"Maximum patient age","title":"Max Age"},"description":"Maximum patient age"},{"name":"gender","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Filter by gender (Male/Female)","title":"Gender"},"description":"Filter by gender (Male/Female)"},{"name":"insurance","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Filter by insurance provider","title":"Insurance"},"description":"Filter by insurance provider"},{"name":"sort_by","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Sort by field","default":"Date of Admission","title":"Sort By"},"description":"Sort by field"},{"name":"order","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Sort order (asc/desc)","default":"desc","title":"Order"},"description":"Sort order (asc/desc)"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/encounters/{encounter_id}":{"get":{"summary":"Get Encounter By Id","description":"Get a single encounter by index ID","operationId":"get_encounter_by_id_api_encounters__encounter_id__get","parameters":[{"name":"encounter_id","in":"path","required":true,"schema":{"type":"integer","title":"Encounter Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/patients":{"get":{"summary":"Get Patients","description":"Get unique patients with their demographics","operationId":"get_patients_api_patients_get","parameters":[{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"default":10,"title":"Limit"}},{"name":"age_group","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Age group (0-17, 18-30, 31-50, 51-70, 70+)","title":"Age Group"},"description":"Age group (0-17, 18-30, 31-50, 51-70, 70+)"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/doctors":{"get":{"summary":"Get Doctors","description":"Get list of doctors with their statistics","operationId":"get_doctors_api_doctors_get","parameters":[{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"default":10,"title":"Limit"}},{"name":"specialty","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Filter by specialty","title":"Specialty"},"description":"Filter by specialty"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/hospitals":{"get":{"summary":"Get Hospitals","description":"Get list of hospitals with their statistics","operationId":"get_hospitals_api_hospitals_get","parameters":[{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"default":10,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/conditions":{"get":{"summary":"Get Conditions","description":"Get all medical conditions with statistics","operationId":"get_conditions_api_conditions_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/medications":{"get":{"summary":"Get Medications","description":"Get all medications with usage statistics","operationId":"get_medications_api_medications_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/insurance":{"get":{"summary":"Get Insurance Providers","description":"Get insurance providers with coverage statistics","operationId":"get_insurance_providers_api_insurance_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/stats":{"get":{"summary":"Get Statistics","description":"Get overall dataset statistics","operationId":"get_statistics_api_stats_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/search":{"get":{"summary":"Search","description":"Search across all fields\nExample: /api/search?q=diabetes","operationId":"search_api_search_get","parameters":[{"name":"q","in":"query","required":true,"schema":{"type":"string","minLength":2,"description":"Search query","title":"Q"},"description":"Search query"},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"default":10,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}